Links to cards now open in overlays
Links to cards - both inside iframes and direct anchor links, now trigger an overlay.

Currently the card ID is passed to the overlay by changing the SRC property of the iframe, but this will be updated so that it passes it as a function all or messagePost event so the frame doesn't need to reload.

Note: This has been coded to gracefully degrade but not yet tested in old versions of IE and are likely issues with it we havent' discovered yet (such as interoperability with flash, forms that might be on a page, etc).